+++ libft.h
@@ -6,12 +6,14 @@
 /*   By: aantonio <aantonio@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/24 14:26:06 by aantonio          #+#    #+#             */
-/*   Updated: 2022/12/29 17:56:52 by aantonio         ###   ########.fr       */
+/*   Updated: 2022/12/30 09:59:04 by aantonio         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef LIBFT_H
 # define LIBFT_H
+
+# include <stddef.h>
 
 int				ft_isalnum(int c);
 int				ft_isalpha(int c);
@@ -25,8 +27,8 @@ int				ft_strncmp(char *s1, char *s2, unsigned int n);
 unsigned long	ft_strlcat(char *dest, char *src, unsigned int destsize);
 unsigned int	ft_strlcpy(char *dest, char *src, unsigned int n);
 void			*ft_memchr(const void *s, int c, unsigned int n);
-void			ft_bzero(void *s, int n);
-void			*ft_memset(void *s, int c, int n);
+void			ft_bzero(void *s, size_t n);
+void			*ft_memset(void *s, int c, size_t n);
 void			ft_memcpy(void *dest, void *src, unsigned int n);
 void			ft_memmove(void *dest, void *src, unsigned int n);
 int				ft_memcmp(const void *s1, const void *s2, unsigned int n);
